feat(variant idx): integrate all components into dag
